+++ packages/yourloops/test/integration/utils/render.tsx
@@ -1,7 +1,7 @@
 import { Router } from 'react-router-dom'
 import { AuthContextProvider } from '../../../lib/auth'
 import { MainLobby } from '../../../app/main-lobby'
-import { createMemoryHistory } from 'history'
+import { createMemoryHistory, MemoryHistory } from 'history'
 import { render } from '@testing-library/react'
 import React from 'react'
 
@@ -15,8 +15,12 @@ function getMainLobby(history) {
   )
 }
 
+export const renderPageFromHistory = (history: MemoryHistory) => {
+  render(getMainLobby(history))
+}
+
 export const renderPage = (url: string) => {
   const history = createMemoryHistory({ initialEntries: [url] })
-  render(getMainLobby(history))
+  renderPageFromHistory(history)
   expect(history.location.pathname).toBe(url)
 }